pyproject.toml
[tool.poetry]
name = "movies-notes-bot"
version = "0.1.0"
description = ""
authors = ["Blissbill ne.rom4nov@gmail.com"]
readme = "README.md"
packages = [{include = "movies_notes_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.1.1"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
sqlalchemy = "^2.0.29"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
